Remove some bloat
